Split `mrt` into several sub-modules
The single file was starting to get unwieldy, and I'm about to add even
more stuff to it.
