--- src/mrt.rs
@@ -1,351 +1,0 @@
-//! API for the MRT (Multi-Rate Timer) peripheral
-//!
-//! Please be aware that this doesn't try to abstract everything, it only
-//! implements the embedded-hal `Timer` functionality.
-//!
-//! The MRT consists of 4 channels, which are mostly separate and can each act
-//! as a run-of-the-mill timer.
-
-use core::convert::TryFrom;
-
-use crate::{
-    pac::{self, mrt0::CHANNEL},
-    reg_proxy::{Reg, RegProxy},
-    syscon,
-};
-
-use embedded_hal::timer::{CountDown, Periodic};
-use embedded_hal_alpha::timer::{
-    CountDown as CountDownAlpha, Periodic as PeriodicAlpha,
-};
-use embedded_time::{
-    clock,
-    duration::{Microseconds, Milliseconds, Nanoseconds, Seconds},
-    fraction::Fraction,
-    Instant,
-};
-use void::Void;
-
-/// Represents the MRT instance
-pub struct MRT {
-    mrt: pac::MRT0,
-}
-
-impl MRT {
-    /// Assumes peripheral is in reset state
-    ///
-    /// This means:
-    /// - Each channel is in repeat mode
-    /// - All channel interrupts are disabled
-    pub(crate) fn new(mrt: pac::MRT0) -> Self {
-        Self { mrt }
-    }
-
-    /// Enables the MRT and splits it into it's four channels
-    pub fn split(self, syscon: &mut syscon::Handle) -> Channels {
-        syscon.enable_clock(&self.mrt);
-
-        Channels::new()
-    }
-
-    /// Return the raw peripheral
-    ///
-    /// This method serves as an escape hatch from the HAL API. It returns the
-    /// raw peripheral, allowing you to do whatever you want with it, without
-    /// limitations imposed by the API.
-    ///
-    /// If you are using this method because a feature you need is missing from
-    /// the HAL API, please [open an issue] or, if an issue for your feature
-    /// request already exists, comment on the existing issue, so we can
-    /// prioritize it accordingly.
-    ///
-    /// [open an issue]: https://github.com/lpc-rs/lpc8xx-hal/issues
-    pub fn free(self) -> pac::MRT0 {
-        self.mrt
-    }
-}
-
-/// The maximum timer value
-pub const MAX_VALUE: Ticks = Ticks(0x7fff_ffff - 1);
-
-/// Represents a MRT0 channel
-///
-/// # `embedded-hal` traits
-/// - [`embedded_hal::timer::CountDown`]
-///
-/// [`embedded_hal::timer::CountDown`]: #impl-CountDown
-pub struct Channel<T: Reg>(RegProxy<T>);
-
-impl<T> Channel<T>
-where
-    T: Trait,
-{
-    fn new() -> Self {
-        Self(RegProxy::new())
-    }
-
-    /// Start the timer
-    ///
-    /// The `reload` argument must be smaller than or equal to [`MAX_VALUE`].
-    ///
-    /// [`MAX_VALUE`]: constant.MAX_VALUE.html
-    pub fn start(&mut self, reload: Ticks) {
-        // This stops the timer, to prevent race conditions when resetting the
-        // interrupt bit
-        self.0.intval.write(|w| {
-            w.load().set_bit();
-            unsafe { w.ivalue().bits(0) }
-        });
-        self.0.stat.write(|w| w.intflag().set_bit());
-        self.0
-            .intval
-            .write(|w| unsafe { w.ivalue().bits(reload.0 + 1) });
-    }
-
-    /// Indicates whether the timer is running
-    pub fn is_running(&self) -> bool {
-        self.0.stat.read().run().is_running()
-    }
-
-    /// Returns the current timer value
-    pub fn value(&self) -> u32 {
-        self.0.timer.read().value().bits()
-    }
-
-    /// Returns the reload value of the timer
-    pub fn reload_value(&self) -> u32 {
-        self.0.intval.read().ivalue().bits()
-    }
-
-    /// Non-blockingly "waits" until the count down finishes
-    fn wait(&mut self) -> nb::Result<(), Void> {
-        if self.0.stat.read().intflag().is_pending_interrupt() {
-            // Reset the interrupt flag
-            self.0.stat.write(|w| w.intflag().set_bit());
-            Ok(())
-        } else {
-            Err(nb::Error::WouldBlock)
-        }
-    }
-}
-
-impl<T> CountDown for Channel<T>
-where
-    T: Trait,
-{
-    /// The timer operates in clock ticks from the system clock, that means it
-    /// runs at 12_000_000 ticks per second if you haven't changed it.
-    ///
-    /// It can also only use values smaller than 0x7FFFFFFF.
-    type Time = Ticks;
-
-    fn start<Time>(&mut self, count: Time)
-    where
-        Time: Into<Self::Time>,
-    {
-        self.start(count.into());
-    }
-
-    fn wait(&mut self) -> nb::Result<(), Void> {
-        self.wait()
-    }
-}
-
-impl<T> CountDownAlpha for Channel<T>
-where
-    T: Trait,
-{
-    type Error = Void;
-
-    /// The timer operates in clock ticks from the system clock, that means it
-    /// runs at 12_000_000 ticks per second if you haven't changed it.
-    ///
-    /// It can also only use values smaller than 0x7FFFFFFF.
-    type Time = Ticks;
-
-    fn try_start<Time>(&mut self, count: Time) -> Result<(), Self::Error>
-    where
-        Time: Into<Self::Time>,
-    {
-        Ok(self.start(count.into()))
-    }
-
-    fn try_wait(&mut self) -> nb::Result<(), Self::Error> {
-        self.wait()
-    }
-}
-
-impl<T> Periodic for Channel<T> where T: Trait {}
-
-impl<T> PeriodicAlpha for Channel<T> where T: Trait {}
-
-impl<T> embedded_time::Clock for Channel<T>
-where
-    T: Trait,
-{
-    type T = u32;
-
-    const SCALING_FACTOR: Fraction = Fraction::new(1, 12_000_000);
-
-    fn try_now(&self) -> Result<Instant<Self>, clock::Error> {
-        if self.is_running() {
-            // embedded-time assumes that clocks are counting up, but we are
-            // counting down here. Thus, the need for some translation.
-            Ok(Instant::new(self.reload_value() - self.value()))
-        } else {
-            Err(clock::Error::NotRunning)
-        }
-    }
-}
-
-/// Represents a number of ticks of the MRT timer
-///
-/// `Ticks` has various `From` and `TryFrom` implementations that provide
-/// integration with `embedded_time` duration types. This not only provides a
-/// more convenient API, it also makes it possible to use the MRT generically,
-/// through the [`CountDown`] trait and a bound like
-/// `Timer::Time: TryFrom<Milliseconds>`, without requiring any knowledge of the
-/// timer's frequency.
-///
-/// However, these conversions have performance implications. For best results,
-/// you should use constants for the original values that you want to convert,
-/// to give the compiler a chance to perform the conversion at compile-time.
-///
-/// [`CountDown`]: embedded_hal::timer::CountDown
-#[derive(Debug, Eq, PartialEq, Ord, PartialOrd)]
-pub struct Ticks(u32);
-
-impl Ticks {
-    /// Creates a `Tick` instance with the given number of ticks
-    ///
-    /// This method is provided as a fallback to avoid performance overhead, in
-    /// case the user knows that `value` fits within `MAX_VALUE`, but the
-    /// compiler can't perform the necessary optimization. Please use any of the
-    /// `From` or `TryFrom` implementations instead, if you can afford it.
-    ///
-    /// # Safety
-    ///
-    /// The user must guarantee that `value <= MAX_VALUE`.
-    pub unsafe fn from_u32(value: u32) -> Self {
-        Self(value)
-    }
-
-    /// Returns the number of ticks of this `Tick` instance.
-    /// You may also use the `Into` implementations instead.
-    pub fn to_u32(&self) -> u32 {
-        self.0
-    }
-}
-
-impl TryFrom<u32> for Ticks {
-    type Error = TickConversionError;
-
-    fn try_from(value: u32) -> Result<Self, Self::Error> {
-        if value > MAX_VALUE.0 {
-            return Err(TickConversionError);
-        }
-
-        Ok(Self(value))
-    }
-}
-
-impl From<Ticks> for u32 {
-    fn from(ticks: Ticks) -> Self {
-        ticks.0
-    }
-}
-
-// Eventually, `Ticks` will need a const-generic argument or something, but as
-// long as everything is hardcoded to 12 MHz, the following will do.
-
-impl From<Nanoseconds> for Ticks {
-    fn from(value: Nanoseconds) -> Self {
-        // This can't possibly fail:
-        // - The multiplication can't overflow after converting to `u64`.
-        // - After the division, the value is guaranteed to fit into the `u32`
-        //   again.
-        // - The maximum possible `value` leads to a result that is smaller than
-        //   `MAX_VALUE`.
-        Self((value.0 as u64 * 12 / 1_000) as u32)
-    }
-}
-
-impl TryFrom<Microseconds> for Ticks {
-    type Error = TickConversionError;
-
-    fn try_from(value: Microseconds) -> Result<Self, Self::Error> {
-        let value = value.0.checked_mul(12).ok_or(TickConversionError)?;
-        Self::try_from(value)
-    }
-}
-
-impl TryFrom<Milliseconds> for Ticks {
-    type Error = TickConversionError;
-
-    fn try_from(value: Milliseconds) -> Result<Self, Self::Error> {
-        let value = value.0.checked_mul(12_000).ok_or(TickConversionError)?;
-        Self::try_from(value)
-    }
-}
-
-impl TryFrom<Seconds> for Ticks {
-    type Error = TickConversionError;
-
-    fn try_from(value: Seconds) -> Result<Self, Self::Error> {
-        let value =
-            value.0.checked_mul(12_000_000).ok_or(TickConversionError)?;
-        Self::try_from(value)
-    }
-}
-
-#[derive(Debug, Eq, PartialEq)]
-/// Indicates that a conversion to [`Tick`] failed
-///
-/// This is the case when the resulting value is larger than [`MAX_VALUE`].
-pub struct TickConversionError;
-
-/// Implemented for types that identify MRT channels
-pub trait Trait: Reg<Target = CHANNEL> + sealed::Sealed {}
-
-macro_rules! channels {
-    ($($channel:ident, $field:ident, $index:expr;)*) => {
-        /// Provides access to the MRT channels
-        pub struct Channels {
-            $(
-                #[allow(missing_docs)]
-                pub $field: Channel<$channel>,
-            )*
-        }
-
-        impl Channels {
-            fn new() -> Self {
-                Self {
-                    $($field: Channel::new(),)*
-                }
-            }
-        }
-
-        $(
-            /// Represents one of the MRT channels
-            ///
-            /// Used as a type parameter for [`Channel`].
-            pub struct $channel;
-
-            reg_cluster_array!($channel, CHANNEL, pac::MRT0, channel, $index);
-
-            impl sealed::Sealed for $channel {}
-            impl Trait for $channel {}
-        )*
-    }
-}
-
-channels!(
-    MRT0, mrt0, 0;
-    MRT1, mrt1, 1;
-    MRT2, mrt2, 2;
-    MRT3, mrt3, 3;
-);
-
-mod sealed {
-    pub trait Sealed {}
-}
